+++ samples/Enum.java
@@ -1,3 +1,0 @@
-public enum Enum {
-    HOT, COLD, CLOUDY, SNOW;
-}