+++ samples/Enum.ts
@@ -1,1 +1,0 @@
-export type Enum = 'HOT' | 'COLD' | 'CLOUDY' | 'SNOW';