+++ test/samples/Enum.ts
@@ -1,0 +1,1 @@
+export type Enum = 'HOT' | 'COLD' | 'CLOUDY' | 'SNOW' | 'UNDER_SCORE';